move matchPath to util
